move network types into network module
