+++ api/build.gradle
@@ -7,6 +7,7 @@ dependencies {
   implementation "com.fasterxml.jackson.module:jackson-module-kotlin"
   implementation "io.javalin:javalin"
   implementation project(":core")
+  implementation project(":p2p")
   implementation project(":jvm-libs:extensions")
 
   testImplementation(testFixtures(project(":core")))
