+++ api/src/main/kotlin/maru/api/node/GetNetworkIdentity.kt
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import io.javalin.http.Context
 import io.javalin.http.Handler
 import io.javalin.http.HttpStatus
-import maru.api.NetworkDataProvider
+import maru.p2p.NetworkDataProvider
 
 /**
  * https://ethereum.github.io/beacon-APIs/#/Node/getNetworkIdentity
