+++ app/src/integrationTest/kotlin/maru/app/MaruConsensusSwitchTest.kt
@@ -31,7 +31,8 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
 import org.web3j.protocol.core.methods.response.EthBlock
 
-class MaruConsensusSwitchTest {
+class
+MaruConsensusSwitchTest {
   companion object {
     private const val VANILLA_EXTRA_DATA_LENGTH = 32
   }
