+++ core/src/testFixtures/kotlin/maru/core/ext/DataGenerators.kt
@@ -24,9 +24,6 @@ import maru.core.Validator
 import maru.executionlayer.manager.ExecutionPayloadStatus
 import maru.executionlayer.manager.ForkChoiceUpdatedResult
 import maru.executionlayer.manager.PayloadStatus
-import maru.p2p.GossipMessageType
-import maru.p2p.Message
-import maru.p2p.Version
 import maru.serialization.rlp.RLPSerializers
 import maru.serialization.rlp.bodyRoot
 import org.apache.tuweni.bytes.Bytes
@@ -154,9 +151,4 @@ object DataGenerators {
 
   fun randomValidPayloadStatus(): PayloadStatus =
     PayloadStatus(ExecutionPayloadStatus.VALID, latestValidHash = Random.nextBytes(32), validationError = null)
-
-  fun randomBlockMessage(blockNumber: ULong = 1uL): Message<SealedBeaconBlock, GossipMessageType> {
-    val sealedBeaconBlock = randomSealedBeaconBlock(blockNumber)
-    return Message(GossipMessageType.BEACON_BLOCK, Version.V1, sealedBeaconBlock)
-  }
 }
